Creating media works but...
For now, it is possible to create Maint Media (NOT NA).
Full <install-creating-uninstall> cycle is implemented.
If any media builder is already installed in system, it will be
uninstalled first. Then new MB will be installed. All ISO of the
installed localization will be created and then the MB will be
uninstalled. After that, new localization of MB will be installed, and
etc.